+++ e2e/cypress/integration/combo-nav.spec.ts
@@ -8,14 +8,14 @@ describe("Test all comboNavs for all styles", () => {
         it(`Test ${style} ${id}`, () => {
             cy.visit(style);
             snapId(id);
-            goCheckNext(id, style);
+            goCheckNext(style, id);
             cy.get(id_input).type("3{enter}");
             snapId(id);
             cy.get(id_input).type("50").blur();
             snapId(id);
             cy.get(id_input).focus().type("{downarrow}{enter}");
             snapId(id);
-            goCheckPrev(id, style);
+            goCheckPrev(style, id);
         });
     }
 });