+++ lib/pagy/extras/foundation.rb
@@ -28,6 +28,7 @@ class Pagy # :nodoc:
 
     # Javascript pagination for foundation: it returns a nav and a JSON tag used by the Pagy.nav javascript
     def pagy_foundation_nav_js(pagy, pagy_id: nil, link_extra: '', **vars)
+      sequels = pagy.sequels(**vars)
       p_id = %( id="#{pagy_id}") if pagy_id
       link = pagy_link_proc(pagy, link_extra: link_extra)
       tags = { 'before' => %(<ul class="pagination">#{pagy_foundation_prev_html pagy, link}),
@@ -36,8 +37,8 @@ class Pagy # :nodoc:
                'gap'    => %(<li class="ellipsis gap" aria-hidden="true"></li>),
                'after'  => %(#{pagy_foundation_next_html pagy, link}</ul>) }
 
-      %(<nav#{p_id} class="pagy-njs pagy-foundation-nav-js" aria-label="Pagination" #{
-        pagy_json_attr(pagy, :nav, tags, (sequels = pagy.sequels(**vars)), pagy.label_sequels(sequels))}></nav>)
+      %(<nav#{p_id} class="#{'pagy-rjs ' if sequels.size > 1}pagy-foundation-nav-js" aria-label="Pagination" #{
+        pagy_json_attr(pagy, :nav, tags, sequels, pagy.label_sequels(sequels))}></nav>)
     end
 
     # Javascript combo pagination for Foundation: it returns a nav and a JSON tag used by the Pagy.combo_nav javascript
@@ -51,7 +52,7 @@ class Pagy # :nodoc:
                     p_pages.to_s.length + 1}rem; padding: 0 0.3rem; margin: 0 0.3rem;">)
 
       %(<nav#{p_id} class="pagy-foundation-combo-nav-js" aria-label="Pagination"><div class="input-group" #{
-          pagy_json_attr pagy, :combo_nav, p_page, pagy_marked_link(link)}>#{
+          pagy_json_attr pagy, :combo_nav, pagy_marked_link(link)}>#{
           if (p_prev  = pagy.prev)
             link.call p_prev, pagy_t('pagy.nav.prev'),
                       'style="margin-bottom: 0" aria-label="previous" class="prev button primary"'