+++ lib/pagy/extras/materialize.rb
@@ -28,6 +28,7 @@ class Pagy # :nodoc:
 
     # Javascript pagination for materialize: it returns a nav and a JSON tag used by the Pagy.nav javascript
     def pagy_materialize_nav_js(pagy, pagy_id: nil, link_extra: '', **vars)
+      sequels = pagy.sequels(**vars)
       p_id = %( id="#{pagy_id}") if pagy_id
       link = pagy_link_proc(pagy, link_extra: link_extra)
 
@@ -37,8 +38,8 @@ class Pagy # :nodoc:
                'gap'    => %(<li class="gap disabled"><a href="#">#{pagy_t 'pagy.nav.gap'}</a></li>),
                'after'  => %(#{pagy_materialize_next_html pagy, link}</ul>) }
 
-      %(<div#{p_id} class="pagy-njs pagy-materialize-nav-js" role="navigation" aria-label="pager" #{
-        pagy_json_attr(pagy, :nav, tags, (sequels = pagy.sequels(**vars)), pagy.label_sequels(sequels))}></div>)
+      %(<div#{p_id} class="#{'pagy-rjs ' if sequels.size > 1}pagy-materialize-nav-js" role="navigation" aria-label="pager" #{
+        pagy_json_attr(pagy, :nav, tags, sequels, pagy.label_sequels(sequels))}></div>)
     end
 
     # Javascript combo pagination for materialize: it returns a nav and a JSON tag used by the Pagy.combo_nav javascript
@@ -53,7 +54,7 @@ class Pagy # :nodoc:
 
       html = %(<ul#{p_id} class="pagy-materialize-combo-nav-js pagination chip" role="navigation")
       %(#{html} aria-label="pager" style="padding-right: 0" #{
-          pagy_json_attr pagy, :combo_nav, p_page, pagy_marked_link(link)}>#{
+          pagy_json_attr pagy, :combo_nav, pagy_marked_link(link)}>#{
           pagy_materialize_prev_html pagy, link, style}<li class="pagy-combo-input">#{
           pagy_t 'pagy.combo_nav_js', page_input: input, count: p_page, pages: p_pages}</li>#{
           pagy_materialize_next_html pagy, link, style}</ul>)